mejoras js
